1225更新/requirements.txt
Flask==1.1.2
gunicorn==19.9.0
line-bot-sdk==1.17.0
psycopg2
